+++ petitparser/lib/src/parser/combinator/not.dart
@@ -1,4 +1,5 @@
 import '../../context/context.dart';
+import '../../context/failure.dart';
 import '../../context/result.dart';
 import '../../core/parser.dart';
 import '../action/pick.dart';
@@ -7,16 +8,16 @@ import '../predicate/any.dart';
 import 'delegate.dart';
 
 extension NotParserExtension<T> on Parser<T> {
-  /// Returns a parser (logical not-predicate) that succeeds whenever the
-  /// receiver fails, but never consumes input.
+  /// Returns a parser (logical not-predicate) that succeeds with the [Failure]
+  /// whenever the receiver fails, but never consumes input.
   ///
   /// For example, the parser `char('_').not().seq(identifier)` accepts
   /// identifiers that do not start with an underscore character. If the parser
   /// `char('_')` accepts the input, the negation and subsequently the
   /// complete parser fails. Otherwise the parser `identifier` is given the
   /// ability to process the complete identifier.
-  Parser<void> not([String message = 'success not expected']) =>
-      NotParser(this, message);
+  Parser<Failure<T>> not([String message = 'success not expected']) =>
+      NotParser<T>(this, message);
 
   /// Returns a parser that consumes any input token (character), but the
   /// receiver.
@@ -25,12 +26,12 @@ extension NotParserExtension<T> on Parser<T> {
   /// The parser fails for inputs like `'a'` or `'Z'`, but succeeds for
   /// input like `'1'`, `'_'` or `'$'`.
   Parser<String> neg([String message = 'input not expected']) =>
-      not(message).seq(any()).pick(1);
+      [not(message), any()].toSequenceParser().pick(1);
 }
 
 /// The not-predicate, a parser that succeeds whenever its delegate does not,
 /// but consumes no input [Parr 1994, 1995].
-class NotParser extends DelegateParser<void> {
+class NotParser<T> extends DelegateParser<Failure<T>> {
   final String message;
 
   NotParser(Parser delegate, this.message)
@@ -38,10 +39,10 @@ class NotParser extends DelegateParser<void> {
         super(delegate);
 
   @override
-  Result<void> parseOn(Context context) {
+  Result<Failure<T>> parseOn(Context context) {
     final result = delegate.parseOn(context);
     if (result.isFailure) {
-      return context.success(null);
+      return context.success(result);
     } else {
       return context.failure(message);
     }
@@ -57,9 +58,9 @@ class NotParser extends DelegateParser<void> {
   String toString() => '${super.toString()}[$message]';
 
   @override
-  NotParser copy() => NotParser(delegate, message);
+  NotParser<T> copy() => NotParser<T>(delegate, message);
 
   @override
-  bool hasEqualProperties(NotParser other) =>
+  bool hasEqualProperties(NotParser<T> other) =>
       super.hasEqualProperties(other) && message == other.message;
 }