+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 * Dart 2.9 compatibility and requirement (in preparation for null-safety).
 * `Success.message` throws an `UnsupportedError`, instead of returning `null`.
 * Make debug tools to produce frame objects with accessors instead of strings.
+* `DelegateParser` has been made abstract to avoid a concrete class in-between abstract classes.
+* `Parser.delegate()` has been removed, use `Parser.settable()` as an equivalent replacement.
+* `Parser.not()` is now returning the failure `Parser<Failure<T>>` as success value, instead of `null`.
 
 ## 3.1.0